+++ rtv/submission_page.py
@@ -1,14 +1,10 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 
-import re
-import time
-
 from . import docs
-from .content import SubmissionContent, SubredditContent
+from .content import SubmissionContent
 from .page import Page, PageController, logged_in
 from .objects import Navigator, Command
-from .exceptions import TemporaryFileError
 
 
 class SubmissionController(PageController):
@@ -16,9 +12,10 @@ class SubmissionController(PageController):
 
 
 class SubmissionPage(Page):
-
     BANNER = docs.BANNER_SUBMISSION
     FOOTER = docs.FOOTER_SUBMISSION
+
+    name = 'submission'
 
     def __init__(self, reddit, term, config, oauth, url=None, submission=None):
         super(SubmissionPage, self).__init__(reddit, term, config, oauth)
@@ -33,27 +30,62 @@ class SubmissionPage(Page):
             self.content = SubmissionContent(
                 submission, term.loader,
                 max_comment_cols=config['max_comment_cols'])
+
         # Start at the submission post, which is indexed as -1
         self.nav = Navigator(self.content.get, page_index=-1)
-        self.selected_subreddit = None
 
-    @SubmissionController.register(Command('SORT_HOT'))
+    def handle_selected_page(self):
+        """
+        Open the subscription page in a subwindow, but close the current page
+        if any other type of page is selected.
+        """
+        if not self.selected_page:
+            pass
+        elif self.selected_page.name == 'subscription':
+            # Launch page in a subwindow
+            self.selected_page = self.selected_page.loop()
+        elif self.selected_page.name in ('subreddit', 'submission', 'inbox'):
+            # Replace the current page
+            self.active = False
+        else:
+            raise RuntimeError(self.selected_page.name)
+
+    def refresh_content(self, order=None, name=None):
+        """
+        Re-download comments and reset the page index
+        """
+        order = order or self.content.order
+        url = name or self.content.name
+
+        # Hack to allow an order specified in the name by prompt_subreddit() to
+        # override the current default
+        if order == 'ignore':
+            order = None
+
+        with self.term.loader('Refreshing page'):
+            self.content = SubmissionContent.from_url(
+                self.reddit, url, self.term.loader, order=order,
+                max_comment_cols=self.config['max_comment_cols'])
+        if not self.term.loader.exception:
+            self.nav = Navigator(self.content.get, page_index=-1)
+
+    @SubmissionController.register(Command('SORT_1'))
     def sort_content_hot(self):
         self.refresh_content(order='hot')
 
-    @SubmissionController.register(Command('SORT_TOP'))
+    @SubmissionController.register(Command('SORT_2'))
     def sort_content_top(self):
         self.refresh_content(order='top')
 
-    @SubmissionController.register(Command('SORT_RISING'))
+    @SubmissionController.register(Command('SORT_3'))
     def sort_content_rising(self):
         self.refresh_content(order='rising')
 
-    @SubmissionController.register(Command('SORT_NEW'))
+    @SubmissionController.register(Command('SORT_4'))
     def sort_content_new(self):
         self.refresh_content(order='new')
 
-    @SubmissionController.register(Command('SORT_CONTROVERSIAL'))
+    @SubmissionController.register(Command('SORT_5'))
     def sort_content_controversial(self):
         self.refresh_content(order='controversial')
 
@@ -62,7 +94,6 @@ class SubmissionPage(Page):
         """
         Toggle the selected comment tree between visible and hidden
         """
-
         current_index = self.nav.absolute_index
         self.content.toggle(current_index)
 
@@ -84,69 +115,26 @@ class SubmissionPage(Page):
         """
         Close the submission and return to the subreddit page
         """
-
         self.active = False
-
-    def refresh_content(self, order=None, name=None):
-        """
-        Re-download comments and reset the page index
-        """
-
-        order = order or self.content.order
-        url = name or self.content.name
-
-        # Hack to allow an order specified in the name by prompt_subreddit() to
-        # override the current default
-        if order == 'ignore':
-            order = None
-
-        with self.term.loader('Refreshing page'):
-            self.content = SubmissionContent.from_url(
-                self.reddit, url, self.term.loader, order=order,
-                max_comment_cols=self.config['max_comment_cols'])
-        if not self.term.loader.exception:
-            self.nav = Navigator(self.content.get, page_index=-1)
-
-    @SubmissionController.register(Command('PROMPT'))
-    def prompt_subreddit(self):
-        """
-        Open a prompt to navigate to a different subreddit
-        """
-
-        name = self.term.prompt_input('Enter page: /')
-        if name is not None:
-            # Check if opening a submission url or a subreddit url
-            # Example patterns for submissions:
-            #     comments/571dw3
-            #     /comments/571dw3
-            #     /r/pics/comments/571dw3/
-            #     https://www.reddit.com/r/pics/comments/571dw3/at_disneyland
-            submission_pattern = re.compile(r'(^|/)comments/(?P<id>.+?)($|/)')
-            match = submission_pattern.search(name)
-            if match:
-                url = 'https://www.reddit.com/comments/{0}'
-                self.refresh_content('ignore', url.format(match.group('id')))
-
-            else:
-                with self.term.loader('Loading page'):
-                    content = SubredditContent.from_name(
-                        self.reddit, name, self.term.loader)
-                if not self.term.loader.exception:
-                    self.selected_subreddit = content
-                    self.active = False
 
     @SubmissionController.register(Command('SUBMISSION_OPEN_IN_BROWSER'))
     def open_link(self):
         """
-        Open the selected item with the web browser
-        """
+        Open the link contained in the selected item.
 
+        If there is more than one link contained in the item, prompt the user
+        to choose which link to open.
+        """
         data = self.get_selected_item()
         if data['type'] == 'Submission':
-            self.term.open_link(data['url_full'])
-            self.config.history.add(data['url_full'])
-        elif data['type'] == 'Comment' and data['permalink']:
-            self.term.open_browser(data['permalink'])
+            link = self.prompt_and_select_link()
+            if link:
+                self.config.history.add(link)
+                self.term.open_link(link)
+        elif data['type'] == 'Comment':
+            link = self.prompt_and_select_link()
+            if link:
+                self.term.open_link(link)
         else:
             self.term.flash()
 
@@ -155,8 +143,10 @@ class SubmissionPage(Page):
         """
         Open the selected item with the system's pager
         """
-
         n_rows, n_cols = self.term.stdscr.getmaxyx()
+
+        if self.config['max_pager_cols'] is not None:
+            n_cols = min(n_cols, self.config['max_pager_cols'])
 
         data = self.get_selected_item()
         if data['type'] == 'Submission':
@@ -173,46 +163,8 @@ class SubmissionPage(Page):
     def add_comment(self):
         """
         Submit a reply to the selected item.
-
-        Selected item:
-            Submission - add a top level comment
-            Comment - add a comment reply
         """
-
-        data = self.get_selected_item()
-        if data['type'] == 'Submission':
-            body = data['text']
-            reply = data['object'].add_comment
-        elif data['type'] == 'Comment':
-            body = data['body']
-            reply = data['object'].reply
-        else:
-            self.term.flash()
-            return
-
-        # Construct the text that will be displayed in the editor file.
-        # The post body will be commented out and added for reference
-        lines = ['  |' + line for line in body.split('\n')]
-        content = '\n'.join(lines)
-        comment_info = docs.COMMENT_FILE.format(
-            author=data['author'],
-            type=data['type'].lower(),
-            content=content)
-
-        with self.term.open_editor(comment_info) as comment:
-            if not comment:
-                self.term.show_notification('Canceled')
-                return
-
-            with self.term.loader('Posting', delay=0):
-                reply(comment)
-                # Give reddit time to process the submission
-                time.sleep(2.0)
-
-            if self.term.loader.exception is None:
-                self.reload_page()
-            else:
-                raise TemporaryFileError()
+        self.reply()
 
     @SubmissionController.register(Command('DELETE'))
     @logged_in
@@ -220,7 +172,6 @@ class SubmissionPage(Page):
         """
         Delete the selected comment
         """
-
         if self.get_selected_item()['type'] == 'Comment':
             self.delete_item()
         else:
@@ -231,7 +182,6 @@ class SubmissionPage(Page):
         """
         Open the selected comment with the URL viewer
         """
-
         data = self.get_selected_item()
         comment = data.get('body') or data.get('text') or data.get('url_full')
         if comment:
@@ -245,7 +195,6 @@ class SubmissionPage(Page):
         Move the cursor up to the comment's parent. If the comment is
         top-level, jump to the previous top-level comment.
         """
-
         cursor = self.nav.absolute_index
         if cursor > 0:
             level = max(self.content.get(cursor)['level'], 1)
@@ -264,7 +213,6 @@ class SubmissionPage(Page):
         Jump to the next comment that's at the same level as the selected
         comment and shares the same parent.
         """
-
         cursor = self.nav.absolute_index
         if cursor >= 0:
             level = self.content.get(cursor)['level']
@@ -340,12 +288,15 @@ class SubmissionPage(Page):
 
             attr = self.term.attr('Created')
             self.term.add_space(win)
-            self.term.add_line(win, '{created}'.format(**data), attr=attr)
+            self.term.add_line(win, '{created}{edited}'.format(**data),
+                               attr=attr)
 
             if data['gold']:
                 attr = self.term.attr('Gold')
                 self.term.add_space(win)
-                self.term.add_line(win, self.term.guilded, attr=attr)
+                count = 'x{}'.format(data['gold']) if data['gold'] > 1 else ''
+                text = self.term.gilded + count
+                self.term.add_line(win, text, attr=attr)
 
             if data['stickied']:
                 attr = self.term.attr('Stickied')
@@ -409,7 +360,8 @@ class SubmissionPage(Page):
 
         attr = self.term.attr('Created')
         self.term.add_space(win)
-        self.term.add_line(win, '{created_long}'.format(**data), attr=attr)
+        self.term.add_line(win, '{created_long}{edited_long}'.format(**data),
+                           attr=attr)
 
         row = len(data['split_title']) + 2
         if data['url_full'] in self.config.history:
@@ -446,7 +398,9 @@ class SubmissionPage(Page):
         if data['gold']:
             attr = self.term.attr('Gold')
             self.term.add_space(win)
-            self.term.add_line(win, self.term.guilded, attr=attr)
+            count = 'x{}'.format(data['gold']) if data['gold'] > 1 else ''
+            text = self.term.gilded + count
+            self.term.add_line(win, text, attr=attr)
 
         if data['nsfw']:
             attr = self.term.attr('NSFW')