+++ tests/test_mime_parsers.py
@@ -27,8 +27,8 @@ URLS = OrderedDict([
         None)),
     ('gfycat', (
         'https://gfycat.com/DeliciousUnfortunateAdouri',
-        'https://giant.gfycat.com/DeliciousUnfortunateAdouri.webm',
-        'video/webm')),
+        'https://giant.gfycat.com/DeliciousUnfortunateAdouri.mp4',
+        'video/mp4')),
     ('youtube', (
         'https://www.youtube.com/watch?v=FjNdYp2gXRY',
         'https://www.youtube.com/watch?v=FjNdYp2gXRY',
@@ -83,7 +83,7 @@ URLS = OrderedDict([
         'video/x-youtube')),
     ('clippituser', (
         'https://www.clippituser.tv/c/edqqld',
-        'https://clips.clippit.tv/edqqld/720.mp4',
+        'https://clips.clippit.tv/edqqld/360.mp4',
         'video/mp4')),
     ('gifs_com', (
         'https://gifs.com/gif/cat-jumps-into-lake-after-guy-sneezes-0VqvRL',